+++ src/econ_attributes.rs
@@ -1,7 +1,7 @@
 //! Includes commonly-used non-enumerated attributes related to economy items. Included for 
 //! convenience.
 
-use crate::{Attribute, AttributeValue, EffectType, DescriptionFormat};
+use crate::{Attribute, AttributeValue, AttributeDef, EffectType, DescriptionFormat};
 
 /// Represents the "is_festivized" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
@@ -9,13 +9,16 @@ pub struct IsFestivized;
 
 impl Attribute for IsFestivized {
     const DEFINDEX: u32 = 2053;
-    const NAME: &str = "is_festivized";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("is_festivized");
-    const DESCRIPTION_STRING: Option<&str> = Some("Festivized");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Unusual;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2053,
+        name: "is_festivized",
+        attribute_class: Some("is_festivized"),
+        description_string: Some("Festivized"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Unusual,
+        hidden: false,
+        stored_as_integer: false,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -34,13 +37,16 @@ pub struct IsAustralium;
 
 impl Attribute for IsAustralium {
     const DEFINDEX: u32 = 2027;
-    const NAME: &str = "is australium item";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("is_australium_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Unusual;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2027,
+        name: "is australium item",
+        attribute_class: Some("is_australium_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Unusual,
+        hidden: true,
+        stored_as_integer: true,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -60,13 +66,16 @@ pub struct KillEater;
 
 impl Attribute for KillEater {
     const DEFINDEX: u32 = 214;
-    const NAME: &str = "kill eater";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("kill_eater");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 214,
+        name: "kill eater",
+        attribute_class: Some("kill_eater"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: true,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -85,13 +94,16 @@ pub struct TradableAfterDate(pub u64);
 
 impl Attribute for TradableAfterDate {
     const DEFINDEX: u32 = 211;
-    const NAME: &str = "tradable after date";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("tradable_after_date");
-    const DESCRIPTION_STRING: Option<&str> = Some("\nTradable After: %s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsDate);
-    const EFFECT_TYPE: EffectType = EffectType::Negative;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 211,
+        name: "tradable after date",
+        attribute_class: Some("tradable_after_date"),
+        description_string: Some("\nTradable After: %s1"),
+        description_format: Some(DescriptionFormat::ValueIsDate),
+        effect_type: EffectType::Negative,
+        hidden: true,
+        stored_as_integer: true,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -110,13 +122,16 @@ pub struct UniqueCraftIndex(pub u64);
 
 impl Attribute for UniqueCraftIndex {
     const DEFINDEX: u32 = 229;
-    const NAME: &str = "unique craft index";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("unique_craft_index");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 229,
+        name: "unique craft index",
+        attribute_class: Some("unique_craft_index"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: true,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -135,13 +150,16 @@ pub struct SupplyCrateSeries(pub u64);
 
 impl Attribute for SupplyCrateSeries {
     const DEFINDEX: u32 = 187;
-    const NAME: &str = "set supply crate series";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("supply_crate_series");
-    const DESCRIPTION_STRING: Option<&str> = Some("Crate Series #%s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 187,
+        name: "set supply crate series",
+        attribute_class: Some("supply_crate_series"),
+        description_string: Some("Crate Series #%s1"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Positive,
+        hidden: false,
+        stored_as_integer: false,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -160,13 +178,16 @@ pub struct CustomNameAttr(pub String);
 
 impl Attribute for CustomNameAttr {
     const DEFINDEX: u32 = 500;
-    const NAME: &str = "custom name attr";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("custom_name_attr");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 500,
+        name: "custom name attr",
+        attribute_class: Some("custom_name_attr"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: false,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -185,13 +206,16 @@ pub struct CustomDescAttr(pub String);
 
 impl Attribute for CustomDescAttr {
     const DEFINDEX: u32 = 501;
-    const NAME: &str = "custom desc attr";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("custom_desc_attr");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 501,
+        name: "custom desc attr",
+        attribute_class: Some("custom_desc_attr"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: false,
+    };
 
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -211,13 +235,16 @@ pub struct GifterAccountId(pub u64);
 
 impl Attribute for GifterAccountId {
     const DEFINDEX: u32 = 186;
-    const NAME: &str = "gifter account id";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("gifter_account_id");
-    const DESCRIPTION_STRING: Option<&str> = Some("\nGift from: %s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAccountId);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 186,
+        name: "gifter account id",
+        attribute_class: Some("gifter_account_id"),
+        description_string: Some("\nGift from: %s1"),
+        description_format: Some(DescriptionFormat::ValueIsAccountId),
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: true,
+    };
 
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -236,13 +263,16 @@ pub struct EventDate(pub u64);
 
 impl Attribute for EventDate {
     const DEFINDEX: u32 = 185;
-    const NAME: &str = "event date";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("event_date");
-    const DESCRIPTION_STRING: Option<&str> = Some("Date Received: %s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsDate);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 185,
+        name: "event date",
+        attribute_class: Some("event_date"),
+        description_string: Some("Date Received: %s1"),
+        description_format: Some(DescriptionFormat::ValueIsDate),
+        effect_type: EffectType::Neutral,
+        hidden: true,
+        stored_as_integer: true,
+    };
 
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -261,13 +291,16 @@ pub struct CustomTextureLo(pub u64);
 
 impl Attribute for CustomTextureLo {
     const DEFINDEX: u32 = 152;
-    const NAME: &str = "custom texture lo";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("custom_texture_lo");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 152,
+        name: "custom texture lo",
+        attribute_class: Some("custom_texture_lo"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: true,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -286,13 +319,16 @@ pub struct CustomTextureHi(pub u64);
 
 impl Attribute for CustomTextureHi {
     const DEFINDEX: u32 = 227;
-    const NAME: &str = "custom texture hi";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("custom_texture_hi");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 227,
+        name: "custom texture hi",
+        attribute_class: Some("custom_texture_hi"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: true,
+    };
 
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -312,13 +348,16 @@ pub struct MakersMarkId(pub u64);
 
 impl Attribute for MakersMarkId {
     const DEFINDEX: u32 = 228;
-    const NAME: &str = "makers mark id";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("makers_mark_id");
-    const DESCRIPTION_STRING: Option<&str> = Some("Crafted by %s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAccountId);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 228,
+        name: "makers mark id",
+        attribute_class: Some("makers_mark_id"),
+        description_string: Some("Crafted by %s1"),
+        description_format: Some(DescriptionFormat::ValueIsAccountId),
+        effect_type: EffectType::Positive,
+        hidden: true,
+        stored_as_integer: true,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {
@@ -337,13 +376,16 @@ pub struct HalloweenVoiceModulation;
 
 impl Attribute for HalloweenVoiceModulation {
     const DEFINDEX: u32 = 1006;
-    const NAME: &str = "SPELL: Halloween voice modulation";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("halloween_voice_modulation");
-    const DESCRIPTION_STRING: Option<&str> = Some("Voices from Below");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 1006,
+        name: "SPELL: Halloween voice modulation",
+        attribute_class: Some("halloween_voice_modulation"),
+        description_string: Some("Voices from Below"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Positive,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
         None
@@ -360,13 +402,16 @@ pub struct HalloweenPumpkinExplosions;
 
 impl Attribute for HalloweenPumpkinExplosions {
     const DEFINDEX: u32 = 1007;
-    const NAME: &str = "SPELL: Halloween pumpkin explosions";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("halloween_pumpkin_explosions");
-    const DESCRIPTION_STRING: Option<&str> = Some("Pumpkin Bombs");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 1007,
+        name: "SPELL: Halloween pumpkin explosions",
+        attribute_class: Some("halloween_pumpkin_explosions"),
+        description_string: Some("Pumpkin Bombs"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Positive,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
         None
@@ -383,13 +428,16 @@ pub struct HalloweenGreenFlames;
 
 impl Attribute for HalloweenGreenFlames {
     const DEFINDEX: u32 = 1008;
-    const NAME: &str = "SPELL: Halloween green flames";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("halloween_green_flames");
-    const DESCRIPTION_STRING: Option<&str> = Some("Halloween Fire");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 1008,
+        name: "SPELL: Halloween green flames",
+        attribute_class: Some("halloween_green_flames"),
+        description_string: Some("Halloween Fire"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Positive,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
         None
@@ -406,13 +454,16 @@ pub struct HalloweenDeathGhosts;
 
 impl Attribute for HalloweenDeathGhosts {
     const DEFINDEX: u32 = 1009;
-    const NAME: &str = "SPELL: Halloween death ghosts";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("halloween_death_ghosts");
-    const DESCRIPTION_STRING: Option<&str> = Some("Exorcism");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 1009,
+        name: "SPELL: Halloween death ghosts",
+        attribute_class: Some("halloween_death_ghosts"),
+        description_string: Some("Exorcism"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Positive,
+        hidden: false,
+        stored_as_integer: false,
+    };
     
     fn attribute_value(&self) -> Option<AttributeValue> {
         None
@@ -425,22 +476,25 @@ impl Attribute for HalloweenDeathGhosts {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem1(pub u64);
+pub struct DynamicRecipeComponentDefinedItem1;
 
 impl Attribute for DynamicRecipeComponentDefinedItem1 {
     const DEFINDEX: u32 = 2000;
-    const NAME: &str = "recipe component defined item 1";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
-
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2000,
+        name: "recipe component defined item 1",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
+    
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -448,22 +502,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem1 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem2(pub u64);
+pub struct DynamicRecipeComponentDefinedItem2;
 
 impl Attribute for DynamicRecipeComponentDefinedItem2 {
     const DEFINDEX: u32 = 2001;
-    const NAME: &str = "recipe component defined item 2";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
-
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2001,
+        name: "recipe component defined item 2",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
+    
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -471,22 +528,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem2 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem3(pub u64);
+pub struct DynamicRecipeComponentDefinedItem3;
 
 impl Attribute for DynamicRecipeComponentDefinedItem3 {
     const DEFINDEX: u32 = 2002;
-    const NAME: &str = "recipe component defined item 3";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2002,
+        name: "recipe component defined item 3",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -494,22 +554,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem3 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem4(pub u64);
+pub struct DynamicRecipeComponentDefinedItem4;
 
 impl Attribute for DynamicRecipeComponentDefinedItem4 {
     const DEFINDEX: u32 = 2003;
-    const NAME: &str = "recipe component defined item 4";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
-
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2003,
+        name: "recipe component defined item 4",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
+    
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -517,22 +580,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem4 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem5(pub u64);
+pub struct DynamicRecipeComponentDefinedItem5;
 
 impl Attribute for DynamicRecipeComponentDefinedItem5 {
     const DEFINDEX: u32 = 2004;
-    const NAME: &str = "recipe component defined item 5";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
-
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2004,
+        name: "recipe component defined item 5",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
+    
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -540,22 +606,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem5 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem6(pub u64);
+pub struct DynamicRecipeComponentDefinedItem6;
 
 impl Attribute for DynamicRecipeComponentDefinedItem6 {
     const DEFINDEX: u32 = 2005;
-    const NAME: &str = "recipe component defined item 6";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
-
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2005,
+        name: "recipe component defined item 6",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
+    
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -563,22 +632,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem6 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem7(pub u64);
+pub struct DynamicRecipeComponentDefinedItem7;
 
 impl Attribute for DynamicRecipeComponentDefinedItem7 {
     const DEFINDEX: u32 = 2006;
-    const NAME: &str = "recipe component defined item 7";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2006,
+        name: "recipe component defined item 7",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -586,22 +658,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem7 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem8(pub u64);
+pub struct DynamicRecipeComponentDefinedItem8;
 
 impl Attribute for DynamicRecipeComponentDefinedItem8 {
     const DEFINDEX: u32 = 2007;
-    const NAME: &str = "recipe component defined item 8";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
-
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2007,
+        name: "recipe component defined item 8",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
+    
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -609,22 +684,25 @@ impl Attribute for DynamicRecipeComponentDefinedItem8 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem9(pub u64);
+pub struct DynamicRecipeComponentDefinedItem9;
 
 impl Attribute for DynamicRecipeComponentDefinedItem9 {
     const DEFINDEX: u32 = 2008;
-    const NAME: &str = "recipe component defined item 9";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2008,
+        name: "recipe component defined item 9",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
-
+    
     fn attribute_float_value(&self) -> Option<f64> {
         None
     }
@@ -632,20 +710,23 @@ impl Attribute for DynamicRecipeComponentDefinedItem9 {
 
 /// Represents the "dynamic_recipe_component_defined_item" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DynamicRecipeComponentDefinedItem10(pub u64);
+pub struct DynamicRecipeComponentDefinedItem10;
 
 impl Attribute for DynamicRecipeComponentDefinedItem10 {
     const DEFINDEX: u32 = 2009;
-    const NAME: &str = "recipe component defined item 10";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("dynamic_recipe_component_defined_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsFromLookupTable);
-    const EFFECT_TYPE: EffectType = EffectType::Neutral;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2009,
+        name: "recipe component defined item 10",
+        attribute_class: Some("dynamic_recipe_component_defined_item"),
+        description_string: None,
+        description_format: Some(DescriptionFormat::ValueIsFromLookupTable),
+        effect_type: EffectType::Neutral,
+        hidden: false,
+        stored_as_integer: false,
+    };
     
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
 
     fn attribute_float_value(&self) -> Option<f64> {
@@ -659,16 +740,19 @@ pub struct ToolTargetItem(pub u64);
 
 impl Attribute for ToolTargetItem {
     const DEFINDEX: u32 = 2012;
-    const NAME: &str = "tool target item";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("tool_target_item");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::ValueIsFromLookupTable;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2012,
+        name: "tool target item",
+        attribute_class: Some("tool_target_item"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::ValueIsFromLookupTable,
+        hidden: true,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
-        Some(self.0.into())
+        None
     }
 
     fn attribute_float_value(&self) -> Option<f64> {
@@ -682,13 +766,16 @@ pub struct SeriesNumber(pub u64);
 
 impl Attribute for SeriesNumber {
     const DEFINDEX: u32 = 2031;
-    const NAME: &str = "series number";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("series_number");
-    const DESCRIPTION_STRING: Option<&str> = None;
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = None;
-    const EFFECT_TYPE: EffectType = EffectType::ValueIsFromLookupTable;
-    const HIDDEN: bool = true;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2031,
+        name: "series number",
+        attribute_class: Some("series_number"),
+        description_string: None,
+        description_format: None,
+        effect_type: EffectType::ValueIsFromLookupTable,
+        hidden: true,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
         Some(self.0.into())
@@ -701,17 +788,20 @@ impl Attribute for SeriesNumber {
 
 /// Represents the "taunt attach particle index" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct TauntAttachParticleIndex(pub u64);
+pub struct SetAttachedParticleTaunt(pub u64);
 
-impl Attribute for TauntAttachParticleIndex {
+impl Attribute for SetAttachedParticleTaunt {
     const DEFINDEX: u32 = 2041;
-    const NAME: &str = "taunt attach particle index";
-    const ATTRIBUTE_CLASS: Option<&str> = None;
-    const DESCRIPTION_STRING: Option<&str> = Some("★ Unusual Effect: %s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsParticleIndex);
-    const EFFECT_TYPE: EffectType = EffectType::Unusual;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = true;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2041,
+        name: "taunt attach particle index",
+        attribute_class: None,
+        description_string: Some("★ Unusual Effect: %s1"),
+        description_format: Some(DescriptionFormat::ValueIsParticleIndex),
+        effect_type: EffectType::Unusual,
+        hidden: false,
+        stored_as_integer: true,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
         Some(self.0.into())
@@ -724,17 +814,20 @@ impl Attribute for TauntAttachParticleIndex {
 
 /// Represents the "set_attached_particle" attribute.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct AttachParticleEffect(pub u64);
+pub struct SetAttachedParticle(pub u64);
 
-impl Attribute for AttachParticleEffect {
+impl Attribute for SetAttachedParticle {
     const DEFINDEX: u32 = 134;
-    const NAME: &str = "attach particle effect";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("set_attached_particle");
-    const DESCRIPTION_STRING: Option<&str> = Some("★ Unusual Effect: %s1");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsParticleIndex);
-    const EFFECT_TYPE: EffectType = EffectType::Unusual;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 134,
+        name: "attach particle effect",
+        attribute_class: Some("set_attached_particle"),
+        description_string: Some("★ Unusual Effect: %s1"),
+        description_format: Some(DescriptionFormat::ValueIsParticleIndex),
+        effect_type: EffectType::Unusual,
+        hidden: false,
+        stored_as_integer: false,
+    };
 
     fn attribute_value(&self) -> Option<AttributeValue> {
         Some(self.0.into())