+++ src/killstreak_tier.rs
@@ -1,4 +1,4 @@
-use crate::{Attribute, AttributeValue, EffectType, DescriptionFormat};
+use crate::{Attribute, AttributeValue, AttributeDef, EffectType, DescriptionFormat};
 use strum::{Display, EnumString, EnumIter, EnumCount};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
@@ -35,13 +35,16 @@ pub enum KillstreakTier {
 /// killstreak_tier
 impl Attribute for KillstreakTier {
     const DEFINDEX: u32 = 2025;
-    const NAME: &str = "killstreak tier";
-    const ATTRIBUTE_CLASS: Option<&str> = Some("killstreak_tier");
-    const DESCRIPTION_STRING: Option<&str> = Some("Killstreaks Active");
-    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
-    const EFFECT_TYPE: EffectType = EffectType::Positive;
-    const HIDDEN: bool = false;
-    const STORED_AS_INTEGER: bool = false;
+    const ATTRIBUTE: AttributeDef = AttributeDef {
+        defindex: 2025,
+        name: "killstreak tier",
+        attribute_class: Some("killstreak_tier"),
+        description_string: Some("Killstreaks Active"),
+        description_format: Some(DescriptionFormat::ValueIsAdditive),
+        effect_type: EffectType::Positive,
+        hidden: false,
+        stored_as_integer: false,
+    };
     
     /// Gets the attribute value.
     fn attribute_value(&self) -> Option<AttributeValue> {